requirements.txt
discord.py==1.3.3
python-aiml==0.9.3 